fix(ivy): fix TestBed overrides for multi-providers
Previous to this change, there was a bug where the "multi" flag would
be stripped from provider overrides, so overrides to multi-providers
would always throw an error that regular providers were mixed in with
multi-providers.

This commit ensures the "multi" flag doesn't get stripped and also
removes logic that causes provider overrides to be added to the providers
array  multiple times.